refactor: remove analyze.go, it is now observer_analyze.go and simplify
